FPIn/MLproject/requirements.txt
pyyaml
numpy
pandas
scikit-learn
datasets
sentence-transformers